+++ packaging/windows/hooks/hook-galacteek.py
@@ -116,12 +116,13 @@ datas += [('galacteek/qml/dweb_lib', f'{pkgrDest}/galacteek/qml/dweb_lib')]
 datas += [('galacteek/ld/sparql/querydb',
            f'{pkgrDest}/galacteek/ld/sparql/querydb')]
 datas += [('galacteek/ui/themes', f'{pkgrDest}/galacteek/ui/themes')]
+datas += [('galacteek/ui/pyramids/binario.zip',
+           f'{pkgrDest}/galacteek/ui/pyramids')]
 datas += [('galacteek/hashmarks', f'{pkgrDest}/galacteek/hashmarks')]
 datas += [('galacteek/docs/manual', f'{pkgrDest}/galacteek/docs/manual')]
 datas += [('galacteek-ld-web4/galacteek_ld_web4',
            f'{pkgrDest}/galacteek_ld_web4')]
 datas += [('packaging/windows/random_username', 'random_username')]
-# datas += [('magic.mgc', '.')]
 
 for root, dirs, files in os.walk('galacteek'):
     for file in files: